Extract related Yast2::Systemd::Unit classes to their own files
