+++ library/systemd/src/lib/yast2/systemd.rb
@@ -1,7 +1,7 @@
 require "yast2/systemd/unit"
+require "yast2/systemd/unit_prop_map"
+require "yast2/systemd/unit_properties"
+require "yast2/systemd/unit_installation_properties"
 require "yast2/systemd/target"
 require "yast2/systemd/service"
 require "yast2/systemd/socket"
-
-# Yast2::Systemd control API
-module Yast2::Systemd; end
