+++ library/systemd/src/lib/yast2/systemd/unit.rb
@@ -1,4 +1,7 @@
 require "yast2/systemctl"
+require "yast2/systemd/unit_prop_map"
+require "yast2/systemd/unit_properties"
+require "yast2/systemd/unit_installation_properties"
 
 require "ostruct"
 
@@ -34,42 +37,7 @@ module Yast2
       Yast.import "Stage"
       include Yast::Logger
 
-      SUPPORTED_TYPES  = %w(service socket target).freeze
-      SUPPORTED_STATES = %w(enabled disabled).freeze
-
-      # Values of {#active_state} fow which we consider a unit "active".
-      #
-      # systemctl.c:check_unit_active uses (active, reloading)
-      # For bsc#884756 we also consider "activating" to be active.
-      # (The remaining states are "deactivating", "inactive", "failed".)
-      #
-      # Yes, depending on systemd states that are NOT covered by their
-      # interface stability promise is fragile.
-      # But: 10 to 50ms per call of systemctl is-active, times 100 to 300 services
-      # (depending on hardware and software installed, VM or not)
-      # is a 1 to 15 second delay (bsc#1045658).
-      # That is why we try hard to avoid many systemctl calls.
-      ACTIVE_STATES = ["active", "activating", "reloading"].freeze
-
-      # A Property Map is a plain Hash(Symbol => String).
-      # It
-      # 1. enumerates the properties we're interested in
-      # 2. maps their Ruby names (snake_case) to systemd names (CamelCase)
-      class PropMap < Hash
-      end
-
-      # @return [PropMap]
-      DEFAULT_PROPMAP = {
-        id:              "Id",
-        pid:             "MainPID",
-        description:     "Description",
-        load_state:      "LoadState",
-        active_state:    "ActiveState",
-        sub_state:       "SubState",
-        unit_file_state: "UnitFileState",
-        path:            "FragmentPath",
-        can_reload:      "CanReload"
-      }.freeze
+      SUPPORTED_TYPES = %w(service socket target).freeze
 
       # with ruby 2.4 delegating ostruct with Forwardable start to write warning
       # so define it manually (bsc#1049433)
@@ -97,24 +65,24 @@ module Yast2
       attr_reader :unit_name
       # @return [String] eg. "service"
       attr_reader :unit_type
-      # @return [PropMap]
+      # @return [Yast2::Systemd::UnitPropMap]
       attr_reader :propmap
       # @return [String]
       #   eg "Failed to get properties: Unit name apache2@.service is not valid."
       attr_reader :error
-      # @return [Properties]
+      # @return [Yast2::Systemd::UnitProperties]
       attr_reader :properties
 
       # @param full_unit_name [String] eg "foo.service"
-      # @param propmap [PropMap]
+      # @param propmap [Yast2::Systemd::UnitPropMap]
       # @param property_text [String,nil]
       #   if provided, use it instead of calling `systemctl show`
-      def initialize(full_unit_name, propmap = {}, property_text = nil)
+      def initialize(full_unit_name, propmap = UnitPropMap.new, property_text = nil)
         @unit_name, dot, @unit_type = full_unit_name.rpartition(".")
         raise "Missing unit type suffix" if dot.empty?
 
         log.warn "Unsupported unit type '#{unit_type}'" unless SUPPORTED_TYPES.include?(unit_type)
-        @propmap = propmap.merge!(DEFAULT_PROPMAP)
+        @propmap = propmap.merge!(UnitPropMap::DEFAULT)
 
         @properties = show(property_text)
         @error = properties.error
@@ -134,10 +102,14 @@ module Yast2
       # @raise [Yast::SystemctlError] if 'systemctl show' cannot be executed
       #
       # @param property_text [String,nil] if provided, use it instead of calling `systemctl show`
-      # @return [Properties]
+      # @return [Yast2::Systemd::UnitProperties]
       def show(property_text = nil)
         # Using different handler during first stage (installation, update, ...)
-        Yast::Stage.initial ? InstallationProperties.new(self) : Properties.new(self, property_text)
+        if Yast::Stage.initial
+          UnitInstallationProperties.new(self)
+        else
+          UnitProperties.new(self, property_text)
+        end
       end
 
       # @raise [Yast::SystemctlError] if the command cannot be executed
@@ -222,134 +194,6 @@ module Yast2
         error << command_result.stderr
         refresh!
         command_result.exit.zero?
-      end
-
-      # Structure holding  properties of systemd unit
-      class Properties < OpenStruct
-        include Yast::Logger
-
-        # @param systemd_unit [Systemd::Unit]
-        # @param property_text [String,nil]
-        #   if provided, use it instead of calling `systemctl show`
-        def initialize(systemd_unit, property_text)
-          super()
-          self[:systemd_unit] = systemd_unit
-          if property_text.nil?
-            raw_output   = load_systemd_properties
-            self[:raw]   = raw_output.stdout
-            self[:error] = raw_output.stderr
-            self[:exit]  = raw_output.exit
-          else
-            self[:raw]   = property_text
-            self[:error] = ""
-            self[:exit]  = 0
-          end
-
-          if !exit.zero? || !error.empty?
-            message = "Failed to get properties for unit '#{systemd_unit.unit_name}' ; "
-            message << "Command `#{raw_output.command}` returned error: #{error}"
-            log.error(message)
-            self[:not_found?] = true
-            return
-          end
-
-          extract_properties
-          self[:active?]     = ACTIVE_STATES.include?(active_state)
-          self[:running?]    = sub_state    == "running"
-          self[:loaded?]     = load_state   == "loaded"
-          self[:not_found?]  = load_state   == "not-found"
-          self[:static?]     = load_state   == "static"
-          self[:enabled?]    = read_enabled_state
-          self[:supported?]  = SUPPORTED_STATES.include?(unit_file_state)
-          self[:can_reload?] = can_reload == "yes"
-        end
-
-      private
-
-        # Check the value of #unit_file_state; its value mirrors UnitFileState dbus property
-        # @return [Boolean] True if enabled, False if not
-        def read_enabled_state
-          # If UnitFileState property is missing (due to e.g. legacy sysvinit service) or
-          # has an unknown entry (e.g. "bad") we must use a different way how to get the
-          # real status of the service.
-          if SUPPORTED_STATES.include?(unit_file_state)
-            state_name_enabled?(unit_file_state)
-          else
-            # Check for exit code of `systemctl is-enabled systemd_unit.name` ; additionally
-            # test the stdout of the command for valid values when the service is enabled
-            # https://www.freedesktop.org/software/systemd/man/systemctl.html#is-enabled%20UNIT%E2%80%A6
-            status = systemd_unit.command("is-enabled")
-            status.exit.zero? && state_name_enabled?(status.stdout)
-          end
-        end
-
-        # Systemd service unit can have various states like enabled, enabled-runtime,
-        # linked, linked-runtime, masked, masked-runtime, static, disabled, invalid.
-        # We test for the return value 'enabled' and 'enabled-runtime' to consider
-        # a service as enabled.
-        # @return [Boolean] True if enabled, False if not
-        def state_name_enabled?(state)
-          ["enabled", "enabled-runtime"].include?(state.strip)
-        end
-
-        def extract_properties
-          systemd_unit.propmap.each do |name, property|
-            self[name] = raw[/#{property}=(.+)/, 1]
-          end
-        end
-
-        def load_systemd_properties
-          names = systemd_unit.propmap.values
-          opts = names.map { |property_name| " --property=#{property_name} " }
-          systemd_unit.command("show", options: opts.join)
-        end
-      end
-
-      # systemd command `systemctl show` is not available during installation
-      # and will return error "Running in chroot, ignoring request." Therefore, we must
-      # avoid calling it in the installation workflow. To keep the API partially
-      # consistent, this class offers a replacement for the Properties above.
-      #
-      # It has two goals:
-      # 1. Checks for existence of the unit based on the stderr from the command
-      #    `systemctl is-enabled`
-      # 2. Retrieves the status enabled|disabled which is needed in the installation
-      #    system. There are currently only 3 commands available for systemd in
-      #    inst-sys/chroot: `systemctl enable|disable|is-enabled`. The rest will return
-      #    the error message mentioned above in this comment.
-      #
-      # Once the inst-sys has running dbus/systemd, this class definition can be removed
-      # together with the condition for Stage.initial in the Systemd::Unit#show.
-      class InstallationProperties < OpenStruct
-        include Yast::Logger
-
-        def initialize(systemd_unit)
-          super()
-          self[:systemd_unit] = systemd_unit
-          self[:status]       = read_status
-          self[:raw]          = status.stdout
-          self[:error]        = status.stderr
-          self[:exit]         = status.exit
-          self[:enabled?]     = status.exit.zero?
-          self[:not_found?]   = service_missing?
-        end
-
-      private
-
-        def read_status
-          systemd_unit.command("is-enabled")
-        end
-
-        # Analyze the exit code and stdout of the command `systemctl is-enabled service_name`
-        # http://www.freedesktop.org/software/systemd/man/systemctl.html#is-enabled%20NAME...
-        def service_missing?
-          # the service exists and it's enabled
-          return false if status.exit.zero?
-          # the service exists and it's disabled
-          return false if status.exit.nonzero? && status.stdout.match(/disabled|masked|linked/)
-          # for all other cases the service does not exist
-          true
-        end
       end
     end
   end
